Expose methods for reading features from GeoJSON and WKT. #194
